+++ functions.go
@@ -6,20 +6,28 @@ package ishell
 type CmdFunc func(args ...string) (output string, err error)
 
 func exitFunc(s *Shell) CmdFunc {
-	return func(args ...string) (output string, err error) {
+	return func(args ...string) (string, error) {
 		s.Stop()
 		return "", nil
 	}
 }
 
 func helpFunc(s *Shell) CmdFunc {
-	return func(args ...string) (output string, err error) {
+	return func(args ...string) (string, error) {
 		s.PrintCommands()
 		return "", nil
+	}
+}
+
+func clearFunc(s *Shell) CmdFunc {
+	return func(args ...string) (string, error) {
+		err := s.ClearScreen()
+		return "", err
 	}
 }
 
 func addDefaultFuncs(s *Shell) {
 	s.Register("exit", exitFunc(s))
 	s.Register("help", helpFunc(s))
+	s.Register("clear", clearFunc(s))
 }